moved cropFormatter func from store to constants folder and removed store/store.js and store/reducer.js
