+++ src/shared/constants.js
@@ -12,6 +12,7 @@ import styled from 'styled-components';
 import moment from 'moment';
 import { Info, MonetizationOn } from '@mui/icons-material';
 import { MapboxApiKey } from './keys';
+import arrayEquals from './functions';
 
 export const ReferenceTooltip = ({
   url, source, type, content, hasLink, title,
@@ -689,3 +690,182 @@ export const callCoverCropApi = async (url) => fetch(url)
     // eslint-disable-next-line no-console
     console.log(err.message);
   });
+
+export const cropDataFormatter = (cropData = [{}]) => {
+  const excludedCropZoneDecisionKeys = ['Exclude', 'Up and Coming', 'Discuss'];
+  // Filter unwanted rows
+  const tjson = cropData.filter((crop) => {
+    if (
+      excludedCropZoneDecisionKeys.includes(crop['Zone Decision'])
+      || crop['Cover Crop Name'] === '__Open Discussion Row'
+    ) {
+      return false;
+    } return true;
+  });
+
+  const formatHalfMonthData = (halfMonthData = []) => {
+    const result = [];
+    let index = 0;
+    while (index < halfMonthData.length) {
+      const timePeriod = {
+        startTime: '',
+        endTime: '',
+        months: [],
+        info: [],
+      };
+      if (timePeriod.months.length === 0) {
+        if (halfMonthData[index].start !== '') {
+          timePeriod.startTime = halfMonthData[index].start;
+          timePeriod.endTime = halfMonthData[index].end;
+          timePeriod.info = [...halfMonthData[index].info];
+        }
+        timePeriod.months.push(halfMonthData[index].month);
+        index += 1;
+      }
+      while (
+        index > 0
+        && index < halfMonthData.length
+        && arrayEquals(halfMonthData[index].info, halfMonthData[index - 1].info)
+      ) {
+        timePeriod.months.push(halfMonthData[index].month);
+        index += 1;
+      }
+      if (timePeriod.months.length > 0) result.push(timePeriod);
+      else index += 1;
+    }
+    return result;
+  };
+
+  const formatTimeToHalfMonthData = (startTime = '', endTime = '', param = '', halfMonthData = []) => {
+    const startIndex = moment(startTime).month() * 2 + (moment(startTime).date() >= 15 ? 1 : 0);
+    const endIndex = moment(endTime).month() * 2 + (moment(endTime).date() >= 15 ? 1 : 0);
+    halfMonthData = halfMonthData.map((data, index) => {
+      if (index >= startIndex && index <= endIndex) {
+        const info = [...data.info, param];
+        let start = '';
+        let end = '';
+        if (data.start === '') start = startTime;
+        else start = moment(data.start).isSameOrBefore(startTime) ? startTime : data.start;
+        if (data.end === '') end = endTime;
+        else end = moment(data.end).isSameOrBefore(endTime) ? data.end : endTime;
+        return {
+          ...data, start, end, info,
+        };
+      }
+      return data;
+    });
+    return halfMonthData;
+  };
+
+  const monthStringBuilder = (vals) => {
+    const params = [
+      'Reliable Establishment',
+      'Freeze/Moisture Risk to Establishment',
+      'Frost Seeding',
+      'Fall/Winter Seeding Rate',
+      'Spring Seeding Rate',
+      'Summer Seeding Rate',
+      'Can Interseed',
+      'Average Frost',
+    ];
+    const val = vals;
+    let halfMonthArr = Array.from({ length: 24 }, (_, i) => ({
+      month: moment().month(Math.floor(i / 2)).format('M'),
+      info: [],
+      start: '',
+      end: '',
+    }));
+
+    params.forEach((param) => {
+      if (val.data['Planting and Growth Windows']?.[`${param}`]) {
+        val.data['Planting and Growth Windows']?.[`${param}`].values.forEach((dateArray) => {
+          const datesArr = dateArray.split('-');
+          // const valStart = moment(datesArr[0], 'YYYY-MM-DD');
+          const valStart = moment(datesArr[0], 'MM/DD/YYYY').format('MM/DD');
+          const valEnd = moment(datesArr[1], 'MM/DD/YYYY').format('MM/DD');
+
+          if (moment(valStart).isSameOrAfter(valEnd)) {
+            const tempStart = '01/01';
+            const tempEnd = '12/31';
+            halfMonthArr = formatTimeToHalfMonthData(valStart, tempEnd, param, halfMonthArr);
+            halfMonthArr = formatTimeToHalfMonthData(tempStart, valEnd, param, halfMonthArr);
+          } else {
+            halfMonthArr = formatTimeToHalfMonthData(valStart, valEnd, param, halfMonthArr);
+          }
+        });
+      }
+    });
+    const halfMonthData = formatHalfMonthData(halfMonthArr);
+    val['Half Month Data'] = halfMonthData;
+
+    // this is temporary, needs to be replaced with wither a fix to calendar growth window component or exporting of json from airtable
+    Object.keys(val).forEach((item) => {
+      if (item.endsWith('Early') || item.endsWith('Mid')) {
+        const uniq = [...new Set(val[item])];
+        const removedOldVals = uniq.filter((u) => !u.endsWith('growth'));
+        val[item] = removedOldVals;
+      }
+    });
+    return val;
+  };
+
+  return tjson.map((crop) => {
+    // remove open discussion row and zone decision !== include
+
+    // let val = { fields: crop };
+    const val = monthStringBuilder(crop);
+
+    val.inBasket = false;
+
+    // if (!val['Nitrogen Fixation']) {
+    //   val['Nitrogen Fixation'] = 0;
+    // }
+
+    // if (!val['Early Spring Growth']) {
+    //   val['Early Spring Growth'] = 0;
+    // }
+
+    // val['Discourages Nematodes'] = val['Disoucrages Nematodes'];
+    // // TODO: do we want the __id value to be apart of the object maybe as altId we need the ID from the API to be unaltered
+    // // val.id = val.__id;
+    // val.Drought = val['Drought Tolerance'];
+    // val.Flood = val['Flood Tolerance'];
+    // val.Heat = val['Heat Tolerance'];
+    // val['Low Fertility'] = val['Low Fertility Tolerance'];
+    // val.Salinity = val['Salinity Tolerance'];
+    // val.Shade = val['Shade Tolerance'];
+    // val['Tillage at Vegetative'] = val['Tillage Termination at Vegetative'];
+    // val['Tillage at Flowering'] = val['Tillage Termination at Flowering'];
+
+    // val['Freezing at Flowering'] = val['Freezing Termination at Flowering'];
+
+    // val['Freezing at Vegetative'] = val['Freezing Termination at Vegetative'];
+    // val['Chemical at Vegetative'] = val['Chemical Termination at Vegetative'];
+    // val['Chemical at Flowering'] = val['Chemical Termination at Flowering'];
+
+    // val['Mow at Flowering'] = val['Mow Termination at Flowering'];
+    // val['Roller Crimp at Flowering'] = val['Roller Crimp Termination at Flowering'];
+
+    // if (!val['Frost Seeding']) {
+    //   val['Frost Seeding'] = false;
+    // } else {
+    //   val['Frost Seeding'] = true;
+    // }
+    // if (!val['Can Aerial Seed']) {
+    //   val['Aerial Seeding'] = false;
+    // } else {
+    //   val['Aerial Seeding'] = true;
+    // }
+
+    // // TODO: not using anymore
+    // if (!val['Pollinator Habitat']) {
+    //   val['Pollinator Habitat'] = 0;
+    // }
+
+    // if (!val['Pollinator Food']) {
+    //   val['Pollinator Food'] = 0;
+    // }
+
+    return val;
+  });
+};
